Move PlayerSettings class to client code
